requirements.txt
flask
flask-cors
requests
beautifulsoup4
newspaper3k
gunicorn
